# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Splitr API
  description: Expense splitting application API
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- Bearer Auth: []
...
paths:
...
  /api/groups/{id}:
    get:
      tags:
      - group-controller
      operationId: getGroupDetails
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupDetailsResponse"
    put:
      tags:
      - group-controller
      operationId: updateGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupResponse"
    delete:
      tags:
      - group-controller
      operationId: deleteGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
  /api/groups:
    get:
      tags:
      - group-controller
      operationId: getUserGroups
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupResponse"
    post:
      tags:
      - group-controller
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupCreateRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupResponse"
  /api/groups/{id}/transfer-admin/{newAdminId}:
    post:
      tags:
      - group-controller
      operationId: transferAdminRole
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: newAdminId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
  /api/groups/{id}/settlements:
    get:
      tags:
      - group-controller
      operationId: listSettlements
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SettlementResponse"
    post:
      tags:
      - group-controller
      operationId: createSettlement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementCreateRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SettlementResponse"
  /api/groups/{id}/members:
    post:
      tags:
      - group-controller
      operationId: addMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
        required: true
      responses:
        "201":
          description: Created
  /api/groups/{id}/leave:
    post:
      tags:
      - group-controller
      operationId: leaveGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
  /api/groups/{id}/balances:
    get:
      tags:
      - group-controller
      operationId: getBalances
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupBalanceResponse"
...
  /api/groups/{id}/settlements/{settlementId}:
    delete:
      tags:
      - group-controller
      operationId: revertSettlement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: settlementId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
  /api/groups/{id}/members/{userId}:
    delete:
      tags:
      - group-controller
      operationId: removeMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
components:
  schemas:
...
    GroupUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        currency:
          type: string
          minLength: 1
        type:
          type: string
          enum:
          - TRIP
          - APARTMENT
          - EVENT
          - OTHER
      required:
      - type
    GroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        currency:
          type: string
        type:
          type: string
          enum:
          - TRIP
          - APARTMENT
          - EVENT
          - OTHER
        balance:
          type: number
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
    GroupCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        currency:
          type: string
          minLength: 1
        type:
          type: string
          enum:
          - TRIP
          - APARTMENT
          - EVENT
          - OTHER
      required:
      - type
    SettlementCreateRequest:
      type: object
      properties:
        payeeId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
        method:
          type: string
        notes:
          type: string
      required:
      - amount
      - payeeId
    SettlementResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        payer:
          $ref: "#/components/schemas/UserSummary"
        payee:
          $ref: "#/components/schemas/UserSummary"
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
        notes:
          type: string
        settledAt:
          type: string
          format: date-time
    AddMemberRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
      required:
      - userId
...
    UserSearchResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
    GroupDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        currency:
          type: string
        type:
          type: string
          enum:
          - TRIP
          - APARTMENT
          - EVENT
          - OTHER
        balance:
          type: number
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMemberResponse"
    GroupMemberResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserSearchResponse"
        role:
          type: string
          enum:
          - ADMIN
          - MEMBER
        joinedAt:
          type: string
          format: date-time
    BalanceEntry:
      type: object
      properties:
        from:
          $ref: "#/components/schemas/UserSummary"
        to:
          $ref: "#/components/schemas/UserSummary"
        amount:
          type: number
    GroupBalanceResponse:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        memberBalances:
          type: array
          items:
            $ref: "#/components/schemas/MemberBalance"
        simplifiedDebts:
          type: array
          items:
            $ref: "#/components/schemas/BalanceEntry"
    MemberBalance:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserSummary"
        balance:
          type: number
...
  securitySchemes:
    Bearer Auth:
      type: http
      name: Bearer Auth
      scheme: bearer
      bearerFormat: JWT